week-1/day-5/requirements.txt
langchain>=1.0,<2.0
langchain-core>=1.0,<2.0
langchain-community>=0.4,<0.5
langchain-text-splitters>=1.0,<2.0
langchain-chroma>=0.2.0
langchain-openai>=0.3.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
python-dotenv>=1.0.1
